Adding 'broker' and 'trader' option, monthly performance, TOB calculator
The broker and trader choice allows you to track trades you made on
different platforms or trades made by others, be it a student or a
mentor. The performance calendar shows the monthly performance. The TOB
calculator is used to calculate the Belgian trading tax on order I place
on IBKR. This is likely useless for most other countries.